`storage`: rename `has_transactional_batches` to `may_have...`
For historical data, we need to consider this value `true` by default-
consider a multi-segment transaction across segments [1] and [2],
where [1] contains the `tx_fence`, data batches, and [2] contains the `commit`
marker. Also assume that for whatever reason, [1] has not been recompacted
since this `has_transactional_batches` marker was added, but [2] has,
meaning [1] has `has_transactional_batches=false` and [2] has it as `true`.

When we compact the log, [2] is compacted and potentially has its `commit`
marker removed while the `tx_fence` in [1] is left unremoved, and the
`raft_data` batches in [1] are _also_ left with the transactional bit set.

We could get into a situation in which the `raft_data` batches are read,
and then some `abort` marker in a segment later in the log is mistakenly
applied, since the `commit` marker is now missing.

Prevent this situation by ensuring that both segments [1] and [2] would
have `may_have_transactional_batches=true` set even if [1] has not been
recompacted since the addition of this marker, leading to compaction of [1]
BEFORE compaction of [2].