raft/tests/raft_fixture_base: eliminate a race when closing a channel
On an attempt to use an invalid channel, it was asynchronously stopped
and then removed from the map. If two fibers do this concurrently one of
them may be still co_awaiting channel::stop() when the other one has
destructed the channel object.

Remove the channel from the map synchronously while the iterator is
still valid and stop it when nothing else can access it. Concurrent
fibers will create (and potentially stop and drop) channels with the
same node id, but they will all be distinct.

This should fix unclosed-gate-destruction and unable-transfer-leadership
errors on the coordinated compaction test.