s/tests: added segment appender test
Added low level test validating segment appender behavior. The test
validates the correctness of an appender by executing operations over
the appender and a simple in memory buffer at the same time. The result
of all operations must be the same i.e. file content that is managed by
the appender must be exactly the same as the in memory iobuf content.

Signed-off-by: Michał Maślanka <michal@redpanda.com>