`rptest`: deflake `LogCompactionTxRemovalTest`
1. An open transaction could hold back LSO in `rm_stm`, and therefore
hold back compaction. Set `transaction.timeout.ms` to a low value and
sleep before waiting on compaction to settle to prevent this from happening.
2. A partition movement during a transaction could result in the transaction
failing, and the producer never reaching the right number of messages
acked. Use `wait_for_acks=False` with `KgoVerifierProducer` to avoid this
situation (and allow some data to be dropped).
3. Set `tolerate_data_loss=True`. This seems to be necessary, unfortunately.