`cluster`: alter `vassert` condition in `reset_with_new_epoch()`
We remove `tx_fence` batches during compaction. It is a valid operation
to update the epoch when not currently in an open transaction.

Without this change, we may hit this `vassert()` when `tx_fence` batches
have been compacted away, and an epoch bump is requested due to applied
transactional batches in `producer_state::apply_data()`.