test/finjector: add delay between failures
Previously the failure injector might lead to a situation in which
Redpanda cluster didn't have any time to recover. This lead to a test
being flaky as most of the tests assume progress. Changed the logic to
introduce an expected delay between failures. This way there will be
time when cluster is healthy and the test logic can continue.

Signed-off-by: Michał Maślanka <michal@redpanda.com>