cl: Improve state transition handling
Make state changing method private and require task implementations to
return which state they wish the task to transition to post run.  This
ensures that the task does not inadvertantly change state when it
shouldn't, for example if the task is stopped it shouldn't transition
back to active.

This addresses a bug where the source topic sync task would be stopped
while the task was running and upon task completion, would transition
back to active, causing the task to become 'stuck' in an active state
while not actually running.

Signed-off-by: Michael Boquard <michael@redpanda.com>