treewide: ruff E731: do not assign a lambda
Do not assign a lambda expression, use a definition instead.

Ran: ruff check --select E731 --fix --unsafe-fixes tests

Details about the rule:
https://docs.astral.sh/ruff/rules/lambda-assignment/

Technically this is an unsafe fix, but the unsafeness is very obscure
as far as I can tell (the doc doesn't even mention the unsafeness), e.g.
if someone introspected the specific type of the function object or
its string representation.

I didn't exhaustively check each case but I looked at a few and there
is nothing like that and it seems unlikely to me (and likely to fail
loudly if it happens).

In one place, I had to change the logic slightly as we were actually
assigning to the variable in different branches. In that case I
changed the variable to be initialized to None first, and then
assigned in each branch.