cl/test: rnot: test compaction while syncing
Produce with a fixed key set and also produce some tombstones to
exercise compaction and tomstone syncing.

When using compaction, replication progress is completed when the max
consumed offsets have converged across the producer and consumer, since,
due to compaction, the number of messages consumed on the source and
target consumer could be different.