tuner: Handle ssh timeouts in tuner tests
We were seeing test failures in GCP because of the tuner ssh commands
timing out.

This seems very reproducible now presumably after my changes to speed up
the tuner.

After investigation it seems that this is caused by RX/TX queue count
lowering which apparently seems to somehow kill the SSH connection and
leaves it in bad state.

This doesn't seem to happen on AWS but only on GCP.

A workaround seems to be to disable verbose logging. However that is
very important to debug any other failures and hence we'd rather keep it
on.

This change instead adds a two second timeout to the ssh commands (the
tuner should run in lower XXXms) which we just ignore and just wait for.
The tuner command still succeeds and tunes correctly without issues and
we check that after the timeout.