sst: introduce footer
An SST structure is a series of data blocks (with the two more blocks being
special), then a footer. The footer contains pointers to these last two
blocks:

The metaindex block this contains meta information like the bloom
filter.
Index block: this block contains an index of each data blocks and the end
key for each block.