lsm/core: add two level iterator
This is an iterator that uses an index to find the correct iterator to
open the right data iterator for. The index is also an iterator. This is
a generic utility that is used both within an SST to open the right
block, as well as within a (non-L0) level within the LSM tree to find the right
file to open.

LevelDB Links:
* https://github.com/google/leveldb/blob/main/table/two_level_iterator.cc