dt: Reduce repeater worker count in MPT
MPT is failing during the restart tests as the kgo-repeater OOMs during
the restarts.

In steady state we use about 95% or so of available memory. Half the
worker count to reduce memory pressure.

This previously wasn't an issue when we were still on i3en which has 8GB
per core but became an issue when we switched to m6id with only 4GB per
core. At the time of the switch the issue was hidden by the failure
fixed in the previous commit.