tx/rm_stm: do not compute LSO when reset is in progress
When a reset is in progress (eg: raft snapshot), the stm may clear all
the inflight transactions and reset the next apply offset. Due to
scheduling points, if an LSO request comes in racily, it may not see a
consistent state of the stm.

This commit returns the last known lso as a conservative approach until
the reset is finished.