e2e_finjector: propagate failure injection errors to test
Sometimes when node fails to stop timely when injecting terminate
failure type it may happen that the cluster is in weird state as the
node is supposed to be stopped while it is not.

Changed the code to propagate the failure injector error to the test and
finally fail the test itself.

Signed-off-by: Michał Maślanka <michal@redpanda.com>