redpanda: do not destroy scheduling groups
Previously, out custom scheduling groups were created and destroyed on
startup/shutdown. This could cause undefined behaviour if there were
some lingering continuations/timer tasks trying to schedule onto the
destroyed scheduling groups.

While an alternative would be to have a global gate that waits for all
work on each scheduling group to finish, this is infeasible in practice,
because continuations are scheduled from all over the codebase using the
implicit current_scheduling_group(), so it is hard to track down all
usages of a scheduling group, and change all places, including seastar
library functions (such as the tls shutdown code) to use a new global
gate.

Instead, it is simpler and safer to extend the lifetime of the
scheduling groups. We stop destroying them, and let the reactor release
them as it is getting deconstructed.

To avoid multiple instances and metric registration conflicts, we enforce
a singleton pattern with a private constructor. This ensures only one
instance exists per process, used by both production and test code.