admin/rpc: support access the auth result in handlers
The simplest way that doesn't introduce extra build dependencies to pass
through the auth result is to use `std::any` and an map container to
be able to pass through arbitrary values from the context.

In the future there might be additional information in the context, like
possibly the proxy information or other global information that could
save from passing into every admin service.