kafka: make metadata return term 0 when missing info
This is a functional revert of
https://github.com/redpanda-data/redpanda/commit/a441290876ff9bb01d6d58532f388b94f3719bfe,
making the handler return term_id(0) instead of -1 when there is no term
metadata for a given partition. The Java client, for instance, would
treat -1 as a signal that this broker doesn't have reliable metadata,
and would drop cached epochs as a result[1], which interferes with
truncation detection from KIP-320.

Note that returning -1 was previously done as a fix for flakiness of
PartitionBalancerTest.

This commit also updates the metadata handler to return an error:
- In the Java client, the term_id(0) is still used, but the error is
  treated as a signal to request another update immediately[2].
- Franz go skips processing the partition altogether if there is an
  error, regardless of whether there is a missing term, opting to retry
  later[3].

Note that the lack of error handling in addition to returning -1 is the
cause of flakiness of WriteCachingFailureInjectionE2ETest.

With this commit, both PartitionBalancerTest and
WriteCachingFailureInjectionE2ETest pass reliably.

Relevant Redpanda commits in reverse chronological order:
- https://github.com/redpanda-data/redpanda/commit/a441290876ff9bb01d6d58532f388b94f3719bfe
- https://github.com/redpanda-data/redpanda/commit/7a60b758ad9ff9f2a2d7ec343f2a404ffaf51e89
- https://github.com/redpanda-data/redpanda/commit/86d583b486d033377be4c9f2b331be25ff6d9632

1. https://github.com/apache/kafka/blob/9529003fffd93a9d7e3f6ff7ab081ed84942fd13/clients/src/main/java/org/apache/kafka/clients/Metadata.java#L596-L601
2. https://github.com/apache/kafka/blob/9529003fffd93a9d7e3f6ff7ab081ed84942fd13/clients/src/main/java/org/apache/kafka/clients/Metadata.java#L519-L528
3. https://github.com/twmb/franz-go/blob/8268a5d078c01d29ca0daa1748fac264e0fc2f11/pkg/kgo/metadata.go#L1011