rptest/ruff: enable ruff check in pre-commit
Enable the ruff linter in pre-commit. It has many useful checks and
rptest is currently clean except for the two exceptions in .ruff.toml.

We restrict it to tests subdir currently as outside of tests there are
still some unclean files, especially in tools.

We use --fix, so that safe fixes will be applied automatically and will
appear in your working directory, ready for staging. Issues which can't
be safely fixed will show a diff and one would then fix them by hand.

Unfortunately, we can't emit the diff for fixes that were able to be
fixed, which is annoying, see:

https://github.com/astral-sh/ruff/issues/21613