`ct/l1`: vectorize `cleaned_ranges` in `compaction_state_update`
Imagine the log is:

```
   Dirty   Clean    Dirty
[ [0, 4], [5, 15], [16, 20] ]
```

In the current `compaction_state_update` implementation, which contains
a single `std::optional<cleaned_range>`, it isn't possible to represent
a compaction job which indexes both dirty ranges `[0, 4]` and `[16, 20]`
and wants to commit a compaction update marking them both as clean.

Vectorize the update in the relevant `serde` structs as well as update
the backend `metastore` update implementations so that representing
the above update is simple.