rptest: skip storage init checks in test
The ClusterBootstrapFiveNodes/test_topic_creation_during_bootstrap test
explicitly runs operations, whose side effects include creation of
consumer offsets topic and test topic, concurrently with cluster startup
which means that immediately after bootup a broker won't have a
predictable storage layout that is non-dependent on the test itself.
Therefore we just disable the checks on this one test. IF we needed to
do this a lot we could create a simple protocol for tests to express
customizations to the check.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>