`storage`: add  `has_transaction_data_batches` to `index_state`
We need to be able to ensure that compaction of `segment`s with
transactional raft data occurs (so that their transactional bit is unset)
before it is safe to remove transactional control batches during compaction.

Add a new flag, `has_transaction_data_batches` which defaults to `true`,
and force compaction of `segments` using it. Likely this will only need
to force a single re-compaction of each `segment` to ensure that
transactional raft batches have their transactional bit unset.