`rptest`: disable idempotent producer in compaction test
Idempotency is not required for this test nor does it have anything
to do with any other expectations. However, it can lead to flakes due
to `OUT_OF_ORDER_SEQUENCE_NUMBER` errors.

This is a protocol issue dealing with how we skip sequence checks on
the first request of an unknown producer:

https://github.com/redpanda-data/redpanda/blob/38d9c6c9f6f9d4e4b87ba54f7506d780ba274170/src/v/cluster/rm_stm.cc#L1151

Because produce requests are dispatched in parallel, they may be reordered,
and we can end up in an `OUT_OF_ORDER_SEQUENCE_NUMBER` case.

https://github.com/redpanda-data/redpanda/blob/38d9c6c9f6f9d4e4b87ba54f7506d780ba274170/src/v/cluster/producer_state.cc#L192-L194

In the produce requests shown below, we can imagine that each batch of 79
records are requests 1,2,3,4,5, with seq_ids 0, 79, 158, 237, 316 in that
order. What has appeared to have happened here is that the requests arrived
instead in order 2,3,4,1, leading us to accept 79, 158, 237, and then try
to produce with seq_id=0, leading to an OUT_OF_ORDER_SEQUENCE_NUMBER error,
and an eventual timeout of the producer as it is hard stuck.

Anyways, clearly none of that has to do with compaction, so just disable
idempotency for the producer in this test.