bazel: remove thirdparty workaround for openssl library
By creating a passthrough 'cc_library' directly from the foreign_cc target
(instead of the dummy cc_library in thirdparty), we now have a proper target
that exposes all lib headers, and we can use that as a dependency.
This satisfies the layering check.

The same logic can potentially be applied to all other `src/v/thirdparty` libraries.