ct: use top-level last applied offset as snapshot epoch
The partition snapshot used to drive L0 GC uses an epoch to identify a
consistent snapshot of state. The snapshot is taken from the topic table
which contains a last applied offset which could be used as an epoch,
but if no changes are happening to the topic table then the epoch
doesn't move. Instead we use the last applied offset of the top-level
mux state machine at the controller level which will move consistently
along with the topic table, but at a non-zero forward pace (in the worst
case the epoch service moves this forward forcefully).

The result without this change is that if no new cloud topics are
created then GC stalls at a non-moving minimum epoch.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>