dt/archival: Turn off aws-chunked encoding when on GCS
boto3 by default will use request checksumming in some cases, like PUTs
to S3, which causes the content to be transferred with the aws-chunked
Content-Encoding. GCS will store the content verbatim and return it with
the same Content-Encoding, which breaks our code. This happens in the
inventory test when we use boto3 to upload faked inventory reports to
GCS using boto3 and GCS's S3-compatible-but-not-that-compatible API.
Redpanda fails to parse the aws-chunked encoding as JSON.