tests: bump ducktape SHA
Bump ducktape to c3266e0740781cb26b56e74715c0da48ad59c6c7, which pulls
in the changes from:

https://github.com/redpanda-data/ducktape/pull/52

For improvements around tests which time out after teardown. For example
such tests will no longer be reported as "Test passed" in the details
but have a message from DT indicating the type of the failure.

This "fixes" CORE-13453 in the sense that it will no longer report
"test passed" so we should not get more failures triaged to that issue,
though the underlying problems may persist.