ct/l1/lsm: add add_objects state updates
Adds add_objects as state updates to be used by an upcoming
domain_manager implementation that will used the new LSM STM.

The structure of the update is similar but slightly different from those
used by the simple_stm: while the simple implementation would validate
invariants while constructing the state update object, this new
implementation validates invariants while constructing the rows to be
written to a database via the new replicated_database class.

The invariants checked are the same as those checked in the simple
implementation, but are separated into distinct input validation and
state validation for readability.

The expectation will be that when building these state updates, there
will be no concurrent updates to the affected rows (e.g. by locking),
ensuring the invariants being checked do not change while the resulting
rows are being replicated.