base_transport: use optional<output_stream>
Change the output_stream member of base_transport to be an
optional<output_stream>, constructed in do_connect and destroyed in
stop().

This is needed because upstream the default ctor and assignment
operator for output_stream are deprecated. In an earlier we commit we
did a similar transformation for input_stream.

Finally, we also hide the _out member and use the out() accessor to get
to it, so we can assert the optional has a value.

Update users of _out to use out().