ssx: introduce clock and time types
We implement a time library not too dissimlar to absl, except it
integrates with seastar well and has a virtual clock interface for
dependency injection in tests (templating classes for clocks really is a
downer). The duration/time types also use saturating math which is
really nice and prevents the footgun of overflow.